+++ QBTapIt/QBTTaskData.h
@@ -11,7 +11,7 @@
 @interface QBTTaskData : NSObject
 
 @property (nonatomic, strong) NSString* songTitle;
-@property (nonatomic, readwrite) UInt16 trackOrder;
+@property (nonatomic, readwrite) NSUInteger trackOrder;
 
 @property (nonatomic, strong) NSString* tapOnTimeData;
 @property (nonatomic, strong) NSString* tapOffTimeData;
